Extract initAutocomplete to src/lib
