+++ src/handlebars/NostoAutocomplete.ts
@@ -1,4 +1,4 @@
-import { autocomplete } from "../lib/autocomplete"
+import { initAutocomplete } from "../lib/wcAutocompleteHandler"
 import {
   fromHandlebarsTemplate,
   defaultHandlebarsTemplate,
@@ -52,32 +52,11 @@ export class NostoAutocomplete extends HTMLElement {
   }
 
   async connectedCallback() {
-    return initAutocomplete(this)
+    return initAutocomplete(this, {
+      handler: fromHandlebarsTemplate,
+      template: defaultHandlebarsTemplate,
+    })
   }
-}
-
-async function initAutocomplete(element: NostoAutocomplete) {
-  const templateId = element.getAttribute("template")
-  const templateElement = templateId
-    ? document.getElementById(templateId)
-    : element.querySelector<HTMLTemplateElement>("template")
-
-  if (!Object.keys(getConfigFromScript(element)).length) {
-    throw new Error("NostoAutocomplete: Missing required config.")
-  }
-
-  const config = getConfigFromScript(element)
-  return await autocomplete({
-    ...config,
-    render: fromHandlebarsTemplate(
-      templateElement?.innerText ?? defaultHandlebarsTemplate
-    ),
-  })
-}
-
-function getConfigFromScript(element: NostoAutocomplete) {
-  const config = element.querySelector("script[autocomplete-config]")
-  return config ? JSON.parse(config.textContent!) : {}
 }
 
 customElements.define("nosto-autocomplete", NostoAutocomplete)
